backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
numpy==1.26.4
tensorflow==2.18.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0
email-validator==2.3.0
